Wizard-The mind reader
version-1.6 (Beta)

BETA version notice
_____________________
This is an early preview and is neither perfect nor fully featured. This is a glimpse of our direction but may change quite a bit without notice.

You can use it to manage your appointments and more features will be added until we feel ready to have it take over the existing interface.

IMPORTANT: this is the admin side and NOT what the clients will see and use, which isn't changing yet.